change api structure
